+++ src/components/Estamp/Home/components/Stamp.tsx
@@ -1,8 +1,9 @@
 import { CheckBadgeIcon } from '@heroicons/react/24/solid';
 import Image from 'next/image';
 import placeHolderPicture from '@/public/images/pfp-placeholder.svg';
-const Stamp = (props: any) => {
-    const router = props.router;
+import { useRouter } from 'next/router';
+const Stamp = () => {
+    const router = useRouter();
     return (
         <div className="my-8 flex w-4/5 flex-col items-center justify-center text-xl font-bold">
             <div className="relative aspect-square h-auto w-full max-w-full rounded-xl bg-white ring-4 ring-blue-950">
@@ -15,7 +16,7 @@ const Stamp = (props: any) => {
             </div>
             <button
                 className="my-4 flex h-12 w-full items-center justify-center rounded-xl bg-yellow ring-4 ring-yellow/40 transition-all duration-300 ease-in-out hover:ring-16"
-                onClick={() => router.push('/redeem')}
+                onClick={() => router.push('/')}
             >
                 <CheckBadgeIcon className="mx-2 h-8 w-8" />
                 <h1>Redeem Ticket</h1>
